refactor: split example components one-per-file and update docs
Split TransitionDemo (432 lines, 12 components), ContextDemo (241 lines,
8 components), and LazyContextDemo (243 lines, 5 components) into
individual files. Extract shared utilities (sleep) and types (Page).
Add shared context definition files for ContextDemo and LazyContextDemo.
Update CLAUDE.md with new examples structure documentation.
