+++ examples/src/components/GlobalPatchDemo.tsx
@@ -1,0 +1,74 @@
+import { $state, commitUIPatch, startUIPatch } from "yielderact";
+import type { Page } from "../types";
+import { sleep } from "../utils";
+import { Clocks } from "./Clocks";
+import { Navigation } from "./Navigation";
+import { AboutPage, ContactPage, HomePage } from "./PageStubs";
+
+export function* GlobalPatchDemo() {
+  const [page, setPage] = yield* $state<Page>("home");
+  const [isPending, setIsPending] = yield* $state(false);
+  const [log, setLog] = yield* $state<string[]>([]);
+
+  const navigate = async (next: Page) => {
+    setIsPending(true);
+    startUIPatch();
+    try {
+      setLog((prev) => [...prev, `[global] navigating to ${next}...`]);
+      await sleep(5000);
+      setPage(next);
+      setLog((prev) => [...prev, `[global] arrived at ${next}`]);
+    } finally {
+      setIsPending(false);
+      commitUIPatch();
+    }
+  };
+
+  return (
+    <div
+      data-testid="global-patch-demo"
+      style={{ border: "1px solid #ddd", borderRadius: "6px", padding: "1rem" }}
+    >
+      <h4 data-testid="global-patch-heading" style={{ marginTop: 0 }}>
+        Global patch -- entire tree frozen
+      </h4>
+      <p style={{ color: "#555", fontSize: "0.9rem" }}>
+        During navigation the <strong>entire page area</strong> is frozen (including the clock
+        below). State changes are computed but DOM stays unchanged until commit.
+      </p>
+
+      <Clocks />
+
+      <Navigation page={page} isPending={isPending} navigate={navigate} scope="global" />
+
+      <div
+        style={{
+          padding: "0.75rem",
+          background: "#f5f5f5",
+          borderRadius: "4px",
+          minHeight: "80px",
+        }}
+      >
+        <HomePage $shown={page === "home"} />
+        <AboutPage $shown={page === "about"} />
+        <ContactPage $shown={page === "contact"} />
+      </div>
+
+      <pre
+        data-testid="global-patch-log"
+        $shown={!!log.length}
+        style={{
+          marginTop: "0.75rem",
+          fontSize: "0.78rem",
+          background: "#1a1a1a",
+          color: "#cfc",
+          padding: "0.5rem",
+          borderRadius: "4px",
+          overflowX: "auto",
+        }}
+      >
+        {log.join("\n")}
+      </pre>
+    </div>
+  );
+}
