+++ examples/src/components/LazyContextDemo.shared.ts
@@ -1,0 +1,11 @@
+import { createContext } from "yielderact";
+
+export type AppState = {
+  user: { name: string; role: string };
+  count: number;
+};
+
+export const AppCtx = createContext<AppState>({
+  user: { name: "Alice", role: "admin" },
+  count: 0,
+});
