+++ examples/src/components/LiveClock.tsx
@@ -1,0 +1,3 @@
+export function* LiveClock({ tick }: { tick: number }) {
+  return <b data-testid="clock-time">{new Date(tick).toLocaleTimeString("en-US")}</b>;
+}
