+++ examples/src/components/PageStubs.tsx
@@ -1,0 +1,26 @@
+export function* HomePage() {
+  return (
+    <div data-testid="page-home">
+      <h3>Home</h3>
+      <p>...</p>
+    </div>
+  );
+}
+
+export function* AboutPage() {
+  return (
+    <div data-testid="page-about">
+      <h3>About</h3>
+      <p>...</p>
+    </div>
+  );
+}
+
+export function* ContactPage() {
+  return (
+    <div data-testid="page-contact">
+      <h3>Contact</h3>
+      <p>...</p>
+    </div>
+  );
+}
