+++ examples/src/types.ts
@@ -1,0 +1,1 @@
+export type Page = "home" | "about" | "contact";
